ref: extract AbstractTriviaNode for the four identical trivia parse nodes
