+++ src/php/Shared/Parser/Node/WhitespaceNode.php
@@ -4,37 +4,4 @@ declare(strict_types=1);
 
 namespace Phel\Shared\Parser\Node;
 
-use Phel\Lang\SourceLocation;
-
-final readonly class WhitespaceNode implements TriviaNodeInterface
-{
-    public function __construct(
-        private string $code,
-        private SourceLocation $startLocation,
-        private SourceLocation $endLocation,
-    ) {}
-
-    public static function createWithToken(Token $token): self
-    {
-        return new self(
-            $token->getCode(),
-            $token->getStartLocation(),
-            $token->getEndLocation(),
-        );
-    }
-
-    public function getCode(): string
-    {
-        return $this->code;
-    }
-
-    public function getStartLocation(): SourceLocation
-    {
-        return $this->startLocation;
-    }
-
-    public function getEndLocation(): SourceLocation
-    {
-        return $this->endLocation;
-    }
-}
+final readonly class WhitespaceNode extends AbstractTriviaNode {}
